+++ src/views/App/ProvinceView/ProvinceDetailView/ProvinceDetailView.tsx
@@ -7,7 +7,7 @@ import { useParams } from "react-router";
 import { administrativeTypeRepository } from "repositories/administrative-type-repository";
 import { provinceRepository } from "repositories/province-repository";
 import { enumService } from "services/enum-service";
-import { formService } from "services/FormService";
+import { formService } from "services/form-service";
 import "views/App/ProvinceView/ProvinceDetailView/ProvinceDetailView.scss";
 import ProvinceDistrictContentTable from "views/App/ProvinceView/ProvinceDetailView/ProvinceDistrictContentTable/ProvinceDistrictContentTable";
